requirements.txt
fastapi==0.116.1
openai==1.98.0
uvicorn==0.35.0
python-dotenv===1.1.1
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
python-multipart==0.0.20
langchain-openai==0.3.28
# langsmith==0.4.11
Jinja2==3.1.6
langdetect==1.0.9
pyodbc==5.2.0
holidays==0.82
azure-identity==1.20.0
redis-entraid==1.0.0